fix: STASS-82 資産値に整数でない値を入力するとエラーになる不具合を修正

+++ training/TradingTraining-CLI.py
@@ -19,7 +19,7 @@ if __name__ == "__main__":
     start_date = dt.datetime.strptime(start_date_str, '%Y%m%d').date()
     end_date = dt.datetime.now().date() if args.e == None else dt.datetime.strptime(args.e, '%Y%m%d').date()
     lot_volumn = 100 if args.l == None else int(args.l)
-    assets = 10000000 if args.a == None else int(args.a)
+    assets = 10000000 if args.a == None else float(args.a)
 
     # 起動時入力された引数に関する動作確認
     print('code=' + args.code)
